Some general code resturcture now its a little more complex with suspends etc
