+++ constants.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#define MAIN_THREAD_NAME        "mainThread"
+#define IO_THREAD_NAME          "IOthread"
+#define PLUGIN_NAME             "op_ditto"
+#define PSP_KERNEL_MODULE_NAME  "sceKernelLibrary"
+#define COMMANDS_RIGHT_ANALOG   0xB
+#define UART_BAUD_RATE         115200
