+++ defines.h
@@ -1,0 +1,4 @@
+#pragma once
+
+int sceKernelRegisterResumeHandler(int reg, int (*handler)(int unk, void *param), void *param);
+int sceKernelRegisterSuspendHandler(int reg, int (*handler)(int unk, void *param), void *param);
