+++ main.c
@@ -11,31 +11,19 @@
 // written by fLaSh, it can be found here:
 // https://wololo.net/talk/viewtopic.php?f=5&t=11689
 //
-#include <pspctrl.h>
-#include <pspdisplay.h>
-#include <pspdisplay_kernel.h>
-#include <pspkernel.h>
-#include <pspmodulemgr.h>
-#include <psputility_sysparam.h>
-#include "systemctrl.h"
-#include "pspuart.h"
+#include "constants.h"
+#include "defines.h"
+#include "controllerPatchingThread.h"
+#include "ioThread.h"
 
-#define PSP_KERNEL_MODULE_NAME          "sceKernelLibrary"
+PSP_MODULE_INFO(PLUGIN_NAME, 0x1000, 0, 1);
+PSP_MAIN_THREAD_ATTR(0); 
 
-#define COMMANDS_RIGHT_ANALOG           0xB
+// Default right analog to center position
+unsigned char xAxis = 127;
+unsigned char yAxis = 127;
 
-PSP_MODULE_INFO("op_ditto", 0x1000, 0, 1);
-PSP_MAIN_THREAD_ATTR(0);
-  
-int sceKernelRegisterResumeHandler(int reg, int (*handler)(int unk, void *param), void *param);
-int sceKernelRegisterSuspendHandler(int reg, int (*handler)(int unk, void *param), void *param);   
-
-void *hooked_readbuffer_func;
-
-unsigned char xAxis = 50;
-unsigned char yAxis = 20;
-
-static void waitForKernel()
+void waitForKernel()
 {
     /* Wait for the kernel to boot */
     while (sceKernelFindModuleByName(PSP_KERNEL_MODULE_NAME) == NULL)
@@ -44,89 +32,29 @@ static void waitForKernel()
     }
 }
 
-int SIO_thread(SceSize args, void *argp)
-{ 
-    // dont do anything until we can confirm all modules are loaded
-    waitForKernel();    
-
-    // Initialize the psp-uart-library
-    pspUARTInit(115200);
-
-    while (1) {
-        pspUARTWrite(COMMANDS_RIGHT_ANALOG);
-
-        pspUARTWaitForData(10000);
-
-        int recievedDataCount = pspUARTAvailable();
-
-        // If we recieved no data skip and request again
-        if (recievedDataCount == 0) 
-        {
-            continue;
-        } 
-        // If only recieved first byte, wait for second
-        else if (recievedDataCount == 1)
-        {
-            pspUARTWaitForData(10000);
-        }
-        
-        // If still not recieved expected two bytes by this point
-        // reset any data recieved and request again
-        if (pspUARTAvailable() != 2) 
-        {
-            pspUARTResetRingBuffer();
-            continue;
-        }
-        
-        int valueX = pspUARTRead();
-        int valueY = pspUARTRead();
-
-        if (valueX != -1 && valueY != -1) {
-            xAxis = valueX;
-            yAxis = valueY;
-        }
-    }
- 
-    return 0;
-}
-
-s32 sceCtrlReadBufferPositive_patch(SceCtrlData *data, u8 nBufs)
+void startThread(const char *threadName, SceKernelThreadEntry threadFunc, SceSize args, void *argp)
 {
-    int k1 = pspSdkSetK1(0);
+    SceUID thid;
 
-    // create a function ptr to sceCtrlReadBufferPositive() to fill the buttons normally
-    s32 (*hooked_readbuffer_func)(SceCtrlData*, u8) = (s32 (*)(SceCtrlData*, u8))sctrlHENFindFunction("sceController_Service", "sceCtrl", 0x1F803938);
-    hooked_readbuffer_func(data, nBufs);
-
-    data->Rsrv[0] = xAxis;
-    data->Rsrv[1] = yAxis;
-
-    pspSdkSetK1(k1);
-    return 0;
+    thid = sceKernelCreateThread(MAIN_THREAD_NAME, threadFunc, 0x18, 0x500, 0, NULL);
+    if (thid >= 0) {
+        sceKernelStartThread(thid, args, argp);
+    }
 }
 
-int main_thread(SceSize args, void *argp)
-{ 
-    // dont do anything until we can confirm all modules are loaded
-    waitForKernel();    
-
-    // enable analog sampling in the kernel
-    sceCtrlSetSamplingMode(PSP_CTRL_MODE_ANALOG);   
-
-    // grab readbufferpositive from the NID
-    hooked_readbuffer_func = (void *)sctrlHENFindFunction("sceController_Service", "sceCtrl", 0x1F803938); 
-
-    // patch it
-    sctrlHENPatchSyscall(hooked_readbuffer_func, sceCtrlReadBufferPositive_patch);
- 
-    return 0;
+void terminateThread(const char *threadName)
+{
+    SceUID thid;   
+    pspSdkReferThreadStatusByName(threadName, &thid, NULL);
+    
+    sceKernelTerminateThread(thid);
 }
 
-int _ResumeHandler(int unk, void *param)
+int resumeHandler(int unk, void *param)
 {
     SceUID ioThid;
 
-    ioThid = sceKernelCreateThread("IOthread", SIO_thread, 0x18, 0x500, 0, NULL);
+    ioThid = sceKernelCreateThread(IO_THREAD_NAME, ioThread, 0x18, 0x500, 0, NULL);
     if (ioThid >= 0) {
         sceKernelStartThread(ioThid, 0, NULL);
     }
@@ -134,11 +62,15 @@ int _ResumeHandler(int unk, void *param)
   return 0;
 }
 
-int _SuspendHandler(int unk, void *param)
+// UART seems to reset upon suspend and stops working, so we need to terminate it
+// and re-initialize it upon resume.
+int suspendHandler(int unk, void *param)
 {
     SceUID ioThid;   
-    pspSdkReferThreadStatusByName("IOthread", &ioThid, NULL);
+    pspSdkReferThreadStatusByName(IO_THREAD_NAME, &ioThid, NULL);
     
+    // If we call `sceKernelTerminateThread` (which would sounds logical) it will hang the PSP after
+    // the second pause and resume cycle. Just FYI.
     sceKernelTerminateDeleteThread(ioThid);
 
     pspUARTTerminate();
@@ -149,23 +81,12 @@ int _SuspendHandler(int unk, void *param)
 /* Create a user thread */
 int module_start(SceSize args, void *argp)
 { 
-    SceUID thid;
-
-    thid = sceKernelCreateThread("thread", main_thread, 0x18, 0x500, 0, NULL);
-    if (thid >= 0) {
-        sceKernelStartThread(thid, args, argp);
-    }
+    startThread(MAIN_THREAD_NAME, controllerPatchingThread, args, argp);
+    startThread(IO_THREAD_NAME, ioThread, args, argp);
     
-    SceUID ioThid;
+    sceKernelRegisterSuspendHandler(0x1F, suspendHandler, 0);
 
-    ioThid = sceKernelCreateThread("IOthread", SIO_thread, 0x18, 0x500, 0, NULL);
-    if (ioThid >= 0) {
-        sceKernelStartThread(ioThid, args, argp);
-    }
-    
-    sceKernelRegisterSuspendHandler(0x1F, _SuspendHandler, 0);
-
-    sceKernelRegisterResumeHandler(0x1F, _ResumeHandler, 0);
+    sceKernelRegisterResumeHandler(0x1F, resumeHandler, 0);
 
     return 0;
 }
@@ -173,15 +94,8 @@ int module_start(SceSize args, void *argp)
 /* terminate the entire module, so we can return to XMB without freezing */
 int module_stop(SceSize args, void *argp)
 {
-    SceUID thid;   
-    pspSdkReferThreadStatusByName("thread", &thid, NULL);
-    
-    sceKernelTerminateThread(thid);
-    
-    SceUID ioThid;   
-    pspSdkReferThreadStatusByName("IOthread", &ioThid, NULL);
-    
-    sceKernelTerminateThread(ioThid);
+    terminateThread(MAIN_THREAD_NAME);
+    terminateThread(IO_THREAD_NAME);
 
     pspUARTTerminate();
